edit item quantity and total with javacript

+++ cart/views.py
@@ -2,7 +2,7 @@ from django.shortcuts import render, redirect, get_object_or_404
 from .models import CartItem, Cart
 from products.models import ProductVariation
 from django.http import JsonResponse
-
+import json
 
 def cart(request):
     items = CartItem.objects.filter(user=request.user)
@@ -13,6 +13,7 @@ def cart(request):
 
 def add_to_cart(request, product_id):
     if request.method == "POST":
+        cart_total = 0
         # Get the current user's cart
         try:
             cart = Cart.objects.get(user=request.user)
@@ -31,26 +32,40 @@ def add_to_cart(request, product_id):
             except CartItem.DoesNotExist:
                 item = CartItem.objects.create(
                     cart=cart, product_variation=variation, quantity=1, user=request.user)
+                item.total = item.product_variation.price * item.quantity
+                cart_total += item.total
+                cart.total = cart_total
+                cart.save()
+                item.save()
+                
         # Redirect the user back to the product page
         return redirect('view_product', product_id=product_id)
     return redirect('view_product', product_id=product_id)
 
 
 def remove_from_cart(request, cart_item_id):
-    cart_item = get_object_or_404(CartItem, pk=cart_item_id)
-    cart_item.delete()
-    return redirect('cart:cart')
+    if request.method == "POST":
+        print("yes")
+        cart_item = get_object_or_404(CartItem, pk=cart_item_id)
+        cart = cart_item.cart
+        new_cart_total = cart.total - cart_item.total
+        cart.total = new_cart_total
+        cart.save()
+        cart_item.delete()
+        data = {
+            'new_cart_total': new_cart_total
+        }
+        return JsonResponse(data)
 
-def update_cart_item(request):
+def update_cart_item(request, item_id):
     if request.method == 'POST':
-        item_id = request.POST['item_id']
-        new_quantity = request.POST['quantity']
+        new_quantity = json.load(request)["quantity"]
         cart_item = get_object_or_404(CartItem, pk=item_id)
-        cart_item.update_quantity(new_quantity)
+        cart_item.update_quantity(int(new_quantity))
         cart = cart_item.cart
         cart.update_total_price()
         data = {
-            'new_total_price': cart_item.price,
-            'new_cart_total': cart.total_price,
+            'new_item_total': cart_item.total,
+            'new_cart_total': cart.total,
         }
         return JsonResponse(data)
